Optimize the getCompatiability() method to use Rest API.
